+++ src/container/SequentialContainer/Vector.ts
@@ -114,11 +114,11 @@ class Vector<T> extends SequentialContainer<T> {
     );
   }
   front() {
-    if (this.empty()) return undefined;
+    if (!this.length) return undefined;
     return this.vector[0];
   }
   back() {
-    if (this.empty()) return undefined;
+    if (!this.length) return undefined;
     return this.vector[this.length - 1];
   }
   forEach(callback: (element: T, index: number) => void) {